+++ src/main/java/jukebox/musicTicker/DelayHandler.java
@@ -1,4 +1,4 @@
-package jukebox;
+package jukebox.musicTicker;
 
 import net.minecraft.client.audio.MusicTicker.MusicType;
